fix(backend): 批量修复 36 个级联编译错误
【上一轮修复后剩余 36 errors，本轮全部清完】

按错误类型分组：

A. Lombok 缺失（2 errors）
- DocumentController / VideoController：补 @Slf4j 注解 + import

B. SysUser 模型缺字段（6 errors，级联）
- 新增 passwordChangedAt（OffsetDateTime）
- 新增 mustChangePassword（Boolean）

C. 类型错误 short→Integer（8 errors）
- AuthService / PasswordService / SsoUserProvisioning
- 把 (short) 0/1/3 强转去掉，让 int 自动装箱为 Integer

D. 时间类型 LocalDateTime→OffsetDateTime（4 errors）
- PasswordService / SsoUserProvisioning / WebAuthnAuthenticationService
- 全部用 OffsetDateTime.now()

E. SsoUserProvisioning 逻辑（2 errors）
- findByEmployeeNo 返回 SysUser（不是 Optional），去掉 Optional 包装
- 改用 null 判断分支

F. Java API 误用（2 errors）
- VideoListController: OffsetDateTime.toInstant() 去掉多余 ZoneOffset.UTC 参数
- OidcProviderFactory: 删掉 byte[].getBytes() 多余链式调用

G. SLF4J 错误语法（3 errors）
- DocumentService / VideoService / WebhookRetryConsumer
- 把 log.error("msg", error=e) Python 风格改为 Java 风格 log.error("msg", e)

H. BouncyCastle API（3 errors）
- SM2Util: ECParameterSpec → ECDomainParameters（BC crypto.params 包）
- BC 的 ECKeyGenerationParameters / ECPrivateKeyParameters / ECPublicKeyParameters 构造器都要求 ECDomainParameters

I. Lombok boolean getter 命名（1 error）
- WebAuthnController: getResidentKey() → isResidentKey()（Lombok 对 boolean 字段生成 is 前缀）

J. Repository 缺方法 + 类型错（3 errors）
- LeakDetectionRepository: 补 findById(Long) default 方法
- DocumentService: setEncrypted(true) → setEncrypted(1)（Integer 字段 0/1 表示）

K. 内部类引用（1 error）
- SsoAuthenticationFilter: 加 import SsoUserProvisioning.SsoLoginResult

L. VideoResource 缺字段 + 类型错（2 errors）
- VideoResource: 新增 updatedAt（OffsetDateTime, FieldFill.INSERT_UPDATE）
- ClassificationEnforcement: video.setUpdatedAt 改用 OffsetDateTime.now()

+++ backend/src/main/java/com/example/edam/service/AuthService.java
@@ -85,7 +85,7 @@ public class AuthService {
         }
 
         // 5. 重置失败计数
-        user.setFailedLoginCount((short) 0);
+        user.setFailedLoginCount(0);
         userRepository.updateById(user);
 
         // 6. 签发 token
@@ -175,10 +175,10 @@ public class AuthService {
      * - 连续 5 次失败 → 锁定账号 30 分钟
      */
     private void handleFailedLogin(SysUser user) {
-        short failedCount = (short) (user.getFailedLoginCount() + 1);
+        int failedCount = user.getFailedLoginCount() + 1;
         user.setFailedLoginCount(failedCount);
         if (failedCount >= MAX_FAILED_ATTEMPTS) {
-            user.setStatus((short) 3); // 锁定
+            user.setStatus(3); // 锁定
             log.warn("账号锁定: employee_no={} failed_count={}",
                 user.getEmployeeNo(), failedCount);
         }
